+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 ## [0.0.3]
 
-* renamed package to cache_it to be less generic
+* Fixed issue where returning entries would include null values 
+* Fixed issue with updating  values with same keys
+* Added toString methods to CacheEntry and CacheIt
+* Added more unit tests
+
+## [0.0.3]
+
+* Renamed package to cache_it to be less generic
 
 ## [0.0.2]